Comment about streaming.
Just noting to future readers about why we read into a buffer for the test. If you do manage to pipe in a stream, it will likely stall unless you use a tiny image or give it a really big buffer, as the server will be waiting for the client to receive the output so it can drain the write buffer.